requirements/base.txt
aiogram==3.22.0
aiohttp=3.12.15
python-dotenv==1.0.0
SQLAlchemy==2.0.21
asyncpg==0.28.0
python-pptx=1.0.2
pandas=2.3.3
openpyxl=3.1.5
